format

+++ lib/interfaces/include/ButtonInterface.h
@@ -1,10 +1,12 @@
 #ifndef BUTTONINTERFACE_H
 #define BUTTONINTERFACE_H
 
+/* ETL Library */
+#include <etl/singleton.h>
+
 /* External Dependencies */
 #include "SharedFirmwareTypes.h"
 #include <Arduino.h>
-#include <etl/singleton.h>
 #include <stdint.h>
 
 
@@ -29,12 +31,10 @@ struct ButtonState_s
 class ButtonInterface
 {
 public:
-    ButtonInterface(
-        size_t pin,
-        uint8_t debounce_ms = default_button_params::DEBOUNCE_MS,
-        bool active_low = true
-    ) :
-        _pin(pin),
+    ButtonInterface(size_t pin,
+                    uint8_t debounce_ms = default_button_params::DEBOUNCE_MS,
+                    bool active_low = true
+    ) : _pin(pin),
         _debounce_ms(debounce_ms),
         _active_low(active_low)
     {
@@ -42,16 +42,22 @@ public:
     }
 
     void update(unsigned long current_millis);
+
     bool is_pressed();
+
     bool is_released();
+
     bool is_held();
+
     unsigned long get_hold_duration_ms(unsigned long current_millis);
 
 private:
-    const size_t _pin;
+
+    const uint8_t _pin;
     const uint8_t _debounce_ms;
     const bool _active_low;
     ButtonState_s _button_state;
+
 };
 
 #endif // BUTTONINTERFACE_H
